changed user email verification method by using STATUS attribute of user in Cognito
